backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.23
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
cryptography==43.0.3
alembic==1.13.2
pydantic==2.9.2
httpx==0.27.2
pytest==7.4.4
pytest-asyncio==0.23.5
starlette==0.37.2
isort==5.13.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

